Backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
Pillow>=9.0.0
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
rembg>=2.0.0
requests>=2.26.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
onnxruntime>=1.8.0
numpy>=1.21.0